Mostly move methods to item.cpp/itemgen.cpp
