+++ card.cpp
@@ -5,6 +5,7 @@
 #include "elona.hpp"
 #include "i18n.hpp"
 #include "item.hpp"
+#include "itemgen.hpp"
 #include "variables.hpp"
 
 #include <iostream>
