+++ init.cpp
@@ -7,6 +7,7 @@
 #include "config.hpp"
 #include "ctrl_file.hpp"
 #include "elona.hpp"
+#include "enchantment.hpp"
 #include "filesystem.hpp"
 #include "fish.hpp"
 #include "foobar_save.hpp"
@@ -15,6 +16,7 @@
 #include "item.hpp"
 #include "item_db.hpp"
 #include "item_material.hpp"
+#include "itemgen.hpp"
 #include "log.hpp"
 #include "macro.hpp"
 #include "main.hpp"
