+++ talk_unique.cpp
@@ -6,6 +6,7 @@
 #include "i18n.hpp"
 #include "item.hpp"
 #include "item_db.hpp"
+#include "itemgen.hpp"
 #include "macro.hpp"
 #include "variables.hpp"
