+++ variables.hpp
@@ -26,6 +26,15 @@ ELONA_EXTERN(elona_vector1<int> ap);
 ELONA_EXTERN(int wishfilter);
 ELONA_EXTERN(int nooracle);
 
+// enchantment.cpp
+ELONA_EXTERN(elona_vector2<int> encammoref);
+
+// item.hpp/itemgen.hpp
+ELONA_EXTERN(elona_vector1<int> randcolor);
+ELONA_EXTERN(elona_vector1<int> isetbook);
+ELONA_EXTERN(elona_vector1<int> rpsourcelist);
+ELONA_EXTERN(elona_vector1<int> moneybox);
+
 ELONA_EXTERN(elona_vector1<int> _randcolor);
 ELONA_EXTERN(elona_vector1<int> aniref);
 ELONA_EXTERN(elona_vector1<int> anix);
@@ -598,10 +607,6 @@ void damage_insanity(int, int);
 void dmgsp(int, int);
 int efstatusfix(int = 0, int = 0, int = 0, int = 0);
 int eleinfo(int = 0, int = 0);
-int encadd(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
-int encfind(int = 0, int = 0);
-bool encfindspec(int = 0, int = 0);
-int encflt(int = 0, int = 0);
 int eqweaponheavy();
 int eqweaponlight();
 int evfind(int = 0);
@@ -625,20 +630,6 @@ int getworker(int = 0, int = 0);
 int iequiploc(int = 0);
 int imeget();
 int implevel(int = 0);
-int inv_find(int = 0, int = 0);
-int inv_getfreeid(int = 0);
-int inv_getowner(int = 0);
-bool inv_getspace(int);
-int inv_sum(int = 0);
-int inv_weight(int = 0);
-int item_cold(int = 0, int = 0);
-int item_find(int = 0, int = 0, int = 0);
-int item_fire(int = 0, int = 0);
-int item_separate(int);
-int item_stack(int = 0, int = 0, int = 0);
-int itemcreate(int = 0, int = 0, int = 0, int = 0, int = 0);
-int itemfind(int = 0, int = 0, int = 0);
-int itemusingfind(int, bool = false);
 int key_direction();
 int access_item_db(int);
 int get_required_craft_materials();
@@ -649,7 +640,6 @@ int can_evade_trap();
 int try_to_disarm_trap();
 int try_to_perceive_npc(int);
 int do_create_character();
-int do_create_item(int, int, int);
 int initialize_quest_map_crop();
 int initialize_random_nefia_rdtype1();
 int initialize_random_nefia_rdtype4();
@@ -734,9 +724,6 @@ int putcard(int = 0, int = 0);
 int randattb();
 int random_material(int = 0, int = 0);
 int randomele();
-int randomenc(int = 0);
-int randomenclv(int = 0);
-int randomencp(int = 0);
 int randskill();
 int relationbetween(int, int);
 int relocate_chara(int = 0, int = 0, int = 0);
@@ -903,8 +890,6 @@ void draw_emo(int = 0, int = 0, int = 0);
 void drawmenu(int = 0);
 void eatstatus(curse_state_t, int);
 void efllistadd(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
-void egoadd(int = 0, int = 0);
-void encremove(int = 0, int = 0, int = 0);
 void eqrandweaponmage(int = 0);
 void equipinfo(int = 0, int = 0, int = 0);
 void evadd(int = 0, int = 0, int = 0);
@@ -931,15 +916,7 @@ void incognitobegin();
 void incognitoend();
 void initcard(int = 0, int = 0, int = 0);
 std::pair<int, int> inv_getheader(int);
-void item_acid(int = 0, int = 0);
-void item_checkknown(int = 0);
-void item_compress(int);
-void item_copy(int = 0, int = 0);
-void item_delete(int = 0);
-void item_exchange(int = 0, int = 0);
-void item_num(int = 0, int = 0);
 std::vector<int> itemlist(int, int);
-void itemturn(int = 0);
 void key_check(int = 0);
 void keyrelease();
 void initialize_rankn();
@@ -966,21 +943,9 @@ void initialize_item_chip();
 void load_random_name_table();
 void load_random_title_table();
 std::string cheer_up_message(int);
-void initialize_enchantment_data();
 void get_enchantment_description(int, int, int, bool = false);
-void initialize_ego_data();
-void add_enchantments_depending_on_ego();
-void add_enchantment_by_fixed_ego();
-void add_enchantments();
 void initialize_item_material_data();
-void initialize_item_material();
-void determine_item_material();
-void change_item_material();
-void apply_item_material();
-void set_material_specific_attributes();
-void itemname_additional_info();
 void initialize_craft_material_data();
-void get_random_item_id();
 void set_item_info();
 void label_1346();
 void end_dmghp();
@@ -1062,8 +1027,6 @@ void label_1577();
 void label_1578();
 void label_1579();
 void label_1580();
-void label_1581();
-void label_1583();
 void eat_rotten_food();
 void cook();
 void apply_general_eating_effect();
@@ -1380,7 +1343,6 @@ void show_game_score_ranking();
 void proc_event();
 void lenfix(std::string&, int = 0);
 void lovemiracle(int = 0);
-void make_dish(int = 0, int = 0);
 void make_sound(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
 void makecardlist();
 void map_converttile();
@@ -1398,8 +1360,6 @@ void map_randsite(int = 0, int = 0);
 void map_reload(const std::string&);
 void map_setfog(int = 0, int = 0);
 void map_tileset(int = 0);
-void mapitem_cold(int = 0, int = 0);
-void mapitem_fire(int = 0, int = 0);
 void matdelmain(int = 0, int = 0);
 void matgetmain(int = 0, int = 0, int = 0);
 void modcorrupt(int = 0);
@@ -1418,7 +1378,6 @@ void page_load();
 void page_save();
 void press(bool = false);
 void put_questtarget();
-void putenclv(int = 0);
 void refreshspeed(int = 0);
 void remain_make(int = 0, int = 0);
 void removeitem(int, int);
@@ -1454,7 +1413,6 @@ std::string sninn(const std::string&);
 std::string snmagic(const std::string&);
 std::string sntrade(const std::string&);
 std::string sntrainer(const std::string&);
-void sortenc(int = 0);
 void spillblood(int = 0, int = 0, int = 0);
 void spillfrag(int = 0, int = 0, int = 0);
 void stxt(int, const std::string&);
